OrchidCore/src/test/java/com/eden/orchid/api/options/extractors/DoubleOptionExtractorTest.java
package com.eden.orchid.api.options.extractors;

import com.caseyjbrooks.clog.Clog;
import com.eden.orchid.api.converters.DoubleConverter;
import com.eden.orchid.api.converters.StringConverter;
import com.eden.orchid.api.options.annotations.DoubleDefault;
import org.json.JSONObject;
import org.testng.annotations.BeforeMethod;
import org.testng.annotations.Test;

import java.lang.reflect.Field;

import static org.hamcrest.MatcherAssert.*;
import static org.hamcrest.Matchers.*;

@Test(groups = {"unit"})
public class DoubleOptionExtractorTest {

    private StringConverter stringConverter;
    private DoubleConverter doubleConverter;
    private DoubleOptionExtractor underTest;
    private String optionKey;
    private JSONObject optionsObject;

    private Field field;

    @BeforeMethod
    public void testSetup() throws Throwable {
        Clog.setMinPriority(Clog.Priority.FATAL);
        stringConverter = new StringConverter();
        doubleConverter = new DoubleConverter(stringConverter);
        underTest = new DoubleOptionExtractor(doubleConverter);
        optionKey = "optionKey";

        optionsObject = new JSONObject();
    }

// Single Values
//----------------------------------------------------------------------------------------------------------------------

    @Test
    public void testCanHandlePrimitiveValue() throws Throwable {
        field = PrimitiveTestClass.class.getField("emptyDefaultField");

        assertThat(underTest.acceptsClass(double.class), is(true));

        optionsObject.put(optionKey, 11.1);
        assertThat(underTest.getOption(field, optionsObject, optionKey), is(11.1));

        optionsObject.remove(optionKey);
        assertThat(optionsObject.has(optionKey), is(false));

        field = PrimitiveTestClass.class.getField("emptyDefaultField");
        assertThat(underTest.getOption(field, optionsObject, optionKey), is(0.0));

        field = PrimitiveTestClass.class.getField("filledDefaultField");
        assertThat(underTest.getOption(field, optionsObject, optionKey), is(11.1));
    }

    @Test
    public void testCanHandleValueClass() throws Throwable {
        field = ClassTestClass.class.getField("emptyDefaultField");

        assertThat(underTest.acceptsClass(Double.class), is(true));

        optionsObject.put(optionKey, 11.1);
        assertThat(underTest.getOption(field, optionsObject, optionKey), is(new Double(11.1)));

        optionsObject.remove(optionKey);
        assertThat(optionsObject.has(optionKey), is(false));

        field = ClassTestClass.class.getField("emptyDefaultField");
        assertThat(underTest.getOption(field, optionsObject, optionKey), is(new Double(0.0)));

        field = ClassTestClass.class.getField("filledDefaultField");
        assertThat(underTest.getOption(field, optionsObject, optionKey), is(new Double(11.1)));
    }

// testing classes
//----------------------------------------------------------------------------------------------------------------------

    public class PrimitiveTestClass {

        public double emptyDefaultField;

        @DoubleDefault(11.1)
        public double filledDefaultField;
    }

    public class ClassTestClass {

        public Double emptyDefaultField;

        @DoubleDefault(11.1)
        public Double filledDefaultField;
    }

}
